+++ esmaulhusna_muslimbg/example/android/app/src/main/kotlin/cemalkarabulakli/esmaulhusna/esmaulhusna_example/MainActivity.kt
@@ -1,4 +1,4 @@
-package cemalkarabulakli.esmaulhusna.esmaulhusna_example
+package cemalkarabulakli.esmaulhusna_muslimbg.esmaulhusna_example
 
 import io.flutter.embedding.android.FlutterActivity