+++ esmaulhusna_muslimbg/ios/Classes/EsmaulhusnaPlugin.swift
@@ -3,7 +3,7 @@ import UIKit
 
 public class EsmaulhusnaPlugin: NSObject, FlutterPlugin {
   public static func register(with registrar: FlutterPluginRegistrar) {
-    let channel = FlutterMethodChannel(name: "esmaulhusna", binaryMessenger: registrar.messenger())
+    let channel = FlutterMethodChannel(name: "esmaulhusna_muslimbg", binaryMessenger: registrar.messenger())
     let instance = EsmaulhusnaPlugin()
     registrar.addMethodCallDelegate(instance, channel: channel)
   }