+++ esmaulhusna_muslimbg/linux/esmaulhusna_plugin.cc
@@ -1,4 +1,4 @@
-#include "include/esmaulhusna/esmaulhusna_plugin.h"
+#include "include/esmaulhusna_muslimbg/esmaulhusna_plugin.h"
 
 #include <flutter_linux/flutter_linux.h>
 #include <gtk/gtk.h>
@@ -66,7 +66,7 @@ void esmaulhusna_plugin_register_with_registrar(FlPluginRegistrar* registrar) {
   g_autoptr(FlStandardMethodCodec) codec = fl_standard_method_codec_new();
   g_autoptr(FlMethodChannel) channel =
       fl_method_channel_new(fl_plugin_registrar_get_messenger(registrar),
-                            "esmaulhusna",
+                            "esmaulhusna_muslimbg",
                             FL_METHOD_CODEC(codec));
   fl_method_channel_set_method_call_handler(channel, method_call_cb,
                                             g_object_ref(plugin),