+++ esmaulhusna_muslimbg/linux/esmaulhusna_plugin_private.h
@@ -1,6 +1,6 @@
 #include <flutter_linux/flutter_linux.h>
 
-#include "include/esmaulhusna/esmaulhusna_plugin.h"
+#include "include/esmaulhusna_muslimbg/esmaulhusna_plugin.h"
 
 // This file exposes some plugin internals for unit testing. See
 // https://github.com/flutter/flutter/issues/88724 for current limitations